+++ app/src/main/java/pjunhyuk/todaysmoon/OpenWeatherAPITask.java
@@ -5,13 +5,13 @@ import android.os.AsyncTask;
 /**
  * Created by YonseiSIT15_PJH_DT on 2016-07-21.
  */
-public class OpenWeatherAPITask extends AsyncTask<Integer, Void, Weather> {
+public class OpenWeatherAPITask extends AsyncTask<Double, Void, Weather> {
     @Override
-    public Weather doInBackground(Integer...params) {
+    public Weather doInBackground(Double...params) {
         OpenWeatherAPIClient client = new OpenWeatherAPIClient();
 
-        int lat = params[0];
-        int lon = params[1];
+        double lat = params[0];
+        double lon = params[1];
         // API 호출
         Weather w = client.getWeather(lat, lon);
